… util.go