This library seems pretty stable and like it could use some optimization. I found that several methods were not very optimized. So I wrote some benchmarks and did some simple things to improve allocations and bounds checks.

The individual commits are logically grouped by function and discuss the optimizations. At the bottom I've included the results of `benchcmp` comparing benchmarks without any optimizations (baeec21) and with all optimizations. Benchmarks were run with this command:

```
go test -test.run=None -test.bench=. -test.benchmem
```

```
$ benchcmp benchmarks/baeec21.txt benchmarks/1c7b0fe.txt
benchmark                         old ns/op     new ns/op     delta
BenchmarkUUID_MarshalJSON-4       3651          1360          -62.75%
BenchmarkUUID_UnmarshalJSON-4     1769          1731          -2.15%
BenchmarkParse-4                  130           100           -23.08%
BenchmarkNew-4                    3064          1456          -52.48%
BenchmarkUUID_String-4            1692          151           -91.08%
BenchmarkUUID_URN-4               1740          159           -90.86%

benchmark                         old allocs     new allocs     delta
BenchmarkUUID_MarshalJSON-4       18             5              -72.22%
BenchmarkUUID_UnmarshalJSON-4     4              4              +0.00%
BenchmarkParse-4                  1              1              +0.00%
BenchmarkNew-4                    13             2              -84.62%
BenchmarkUUID_String-4            12             1              -91.67%
BenchmarkUUID_URN-4               12             1              -91.67%

benchmark                         old bytes     new bytes     delta
BenchmarkUUID_MarshalJSON-4       712           360           -49.44%
BenchmarkUUID_UnmarshalJSON-4     320           320           +0.00%
BenchmarkParse-4                  16            16            +0.00%
BenchmarkNew-4                    320           64            -80.00%
BenchmarkUUID_String-4            304           48            -84.21%
BenchmarkUUID_URN-4               304           48            -84.21%
```